OneDrive/Desktop/pdf-chatbot/requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
langchain-text-splitters
python-multipart